Refine inline VS Code serve-web launch
