+++ .github/swift-file-length-budget.tsv
@@ -51,8 +51,8 @@
 1945	Sources/RestorableAgentSession.swift
 1900	cmuxTests/NotificationAndMenuBarTests.swift
 1810	Sources/SessionIndexStore.swift
+1749	cmuxTests/OmnibarAndToolsTests.swift
 1748	Sources/WindowDragHandleView.swift
-1744	cmuxTests/OmnibarAndToolsTests.swift
 1732	cmuxTests/WorkspacePullRequestSidebarTests.swift
 1687	cmuxTests/MarkdownPanelTests.swift
 1680	cmuxUITests/BrowserPaneNavigationKeybindUITests.swift
@@ -81,7 +81,6 @@
 1166	Sources/VaultAgentProcessScanner.swift
 1161	cmuxTests/SidebarOrderingTests.swift
 1144	cmuxTests/PiVaultAgentPersistenceTests.swift
-1132	Sources/App/TerminalDirectoryOpenSupport.swift
 1121	cmuxTests/AgentHibernationTests.swift
 1110	Sources/AppDelegate+CmuxSSHURL.swift
 1093	cmuxUITests/BonsplitTabDragUITests.swift
