+++ Sources/App/TerminalDirectoryOpenSupport.swift
@@ -2,14 +2,7 @@ import CmuxFoundation
 import CmuxCore
 import AppKit
 import CmuxCommandPalette
-import Darwin
 import Foundation
-import os
-
-nonisolated private let vscodeServeWebLogger = Logger(
-    subsystem: "com.cmuxterm.app",
-    category: "vscode.serve-web"
-)
 
 enum FinderServicePathResolver {
     private static func canonicalDirectoryPath(_ path: String) -> String {
@@ -307,790 +300,6 @@ enum TerminalDirectoryOpenTarget: String, CaseIterable {
             deduped.append(path)
         }
         return deduped
-    }
-}
-
-enum VSCodeServeWebURLBuilder {
-    static func extractWebUIURL(from output: String) -> URL? {
-        let prefix = "Web UI available at "
-        for line in output.split(whereSeparator: \.isNewline).reversed() {
-            guard let range = line.range(of: prefix) else { continue }
-            let rawURL = line[range.upperBound...].trimmingCharacters(in: .whitespacesAndNewlines)
-            guard !rawURL.isEmpty, let url = URL(string: rawURL) else { continue }
-            return url
-        }
-        return nil
-    }
-
-    static func openFolderURL(baseWebUIURL: URL, directoryPath: String) -> URL? {
-        var components = URLComponents(url: baseWebUIURL, resolvingAgainstBaseURL: false)
-        var queryItems = components?.queryItems ?? []
-        queryItems.removeAll { $0.name == "folder" }
-        queryItems.append(URLQueryItem(name: "folder", value: directoryPath))
-        components?.queryItems = queryItems
-        return components?.url
-    }
-}
-
-nonisolated struct VSCodeCLILaunchConfiguration {
-    let executableURL: URL
-    let argumentsPrefix: [String]
-    let environment: [String: String]
-    let usesCodeTunnelWrapper: Bool
-    let supportsUserDataDirectoryArgument: Bool
-
-    func processArguments(for launchOptions: VSCodeServeWebLaunchOptions) -> [String] {
-        argumentsPrefix + launchOptions.arguments(
-            includeUserDataDirectory: supportsUserDataDirectoryArgument
-        )
-    }
-
-    func processEnvironment(for launchOptions: VSCodeServeWebLaunchOptions) -> [String: String] {
-        var processEnvironment = environment
-        if usesCodeTunnelWrapper {
-            processEnvironment["VSCODE_CLI_USE_FILE_KEYRING"] = "1"
-            if processEnvironment["VSCODE_CLI_DATA_DIR"]?.isEmpty ?? true {
-                processEnvironment["VSCODE_CLI_DATA_DIR"] = launchOptions.serverDataDirectoryURL
-                    .appendingPathComponent("cli-data", isDirectory: true)
-                    .path
-            }
-        }
-        return processEnvironment
-    }
-}
-
-nonisolated struct VSCodeServeWebLaunchOptions: Equatable {
-    static let portEnvironmentKey = "CMUX_VSCODE_SERVE_WEB_PORT"
-    static let dataDirectoryEnvironmentKey = "CMUX_VSCODE_SERVE_WEB_DATA_DIR"
-    static let portDefaultsKey = "vscodeServeWeb.port"
-
-    let port: Int
-    let serverDataDirectoryURL: URL
-    let userDataDirectoryURL: URL
-    let connectionTokenFileURL: URL
-    let allowsEphemeralPortFallback: Bool
-
-    var arguments: [String] {
-        arguments(includeUserDataDirectory: true)
-    }
-
-    func arguments(includeUserDataDirectory: Bool) -> [String] {
-        var arguments = [
-            "--accept-server-license-terms",
-            "--host", "127.0.0.1",
-            "--port", String(port),
-            "--server-data-dir", serverDataDirectoryURL.path,
-        ]
-        if includeUserDataDirectory {
-            arguments.append(contentsOf: ["--user-data-dir", userDataDirectoryURL.path])
-        }
-        arguments.append(contentsOf: [
-            "--connection-token-file", connectionTokenFileURL.path,
-        ])
-        return arguments
-    }
-
-    static func resolve(
-        environment: [String: String] = ProcessInfo.processInfo.environment,
-        defaults: UserDefaults = .standard,
-        fileManager: FileManager = .default,
-        bundleIdentifier: String? = Bundle.main.bundleIdentifier,
-        applicationSupportDirectoryURL: URL? = FileManager.default.urls(
-            for: .applicationSupportDirectory,
-            in: .userDomainMask
-        ).first
-    ) -> VSCodeServeWebLaunchOptions? {
-        guard let serverDataDirectoryURL = resolveServerDataDirectoryURL(
-            environment: environment,
-            bundleIdentifier: bundleIdentifier,
-            applicationSupportDirectoryURL: applicationSupportDirectoryURL
-        ) else { return nil }
-        let userDataDirectoryURL = serverDataDirectoryURL.appendingPathComponent("user-data", isDirectory: true)
-
-        do {
-            try fileManager.createDirectory(
-                at: serverDataDirectoryURL,
-                withIntermediateDirectories: true
-            )
-            try fileManager.createDirectory(
-                at: userDataDirectoryURL,
-                withIntermediateDirectories: true
-            )
-        } catch {
-            vscodeServeWebLogger.error(
-                "Failed to create VS Code serve-web directories: \(error.localizedDescription, privacy: .private)"
-            )
-            return nil
-        }
-
-        guard let connectionTokenFileURL = ensureConnectionTokenFile(
-            in: serverDataDirectoryURL,
-            fileManager: fileManager
-        ) else { return nil }
-
-        let portResolution = resolvePort(
-            environment: environment,
-            defaults: defaults,
-            bundleIdentifier: bundleIdentifier
-        )
-
-        return VSCodeServeWebLaunchOptions(
-            port: portResolution.port,
-            serverDataDirectoryURL: serverDataDirectoryURL,
-            userDataDirectoryURL: userDataDirectoryURL,
-            connectionTokenFileURL: connectionTokenFileURL,
-            allowsEphemeralPortFallback: portResolution.allowsEphemeralFallback
-        )
-    }
-
-    func ephemeralPortFallback() -> VSCodeServeWebLaunchOptions? {
-        guard allowsEphemeralPortFallback else { return nil }
-        return VSCodeServeWebLaunchOptions(
-            port: 0,
-            serverDataDirectoryURL: serverDataDirectoryURL,
-            userDataDirectoryURL: userDataDirectoryURL,
-            connectionTokenFileURL: connectionTokenFileURL,
-            allowsEphemeralPortFallback: false
-        )
-    }
-
-    private static func resolveServerDataDirectoryURL(
-        environment: [String: String],
-        bundleIdentifier: String?,
-        applicationSupportDirectoryURL: URL?
-    ) -> URL? {
-        if let rawPath = environment[dataDirectoryEnvironmentKey]?.trimmingCharacters(in: .whitespacesAndNewlines),
-           !rawPath.isEmpty {
-            return URL(fileURLWithPath: expandedHomePath(rawPath), isDirectory: true)
-        }
-
-        guard let applicationSupportDirectoryURL else { return nil }
-        let namespace = normalizedNamespace(bundleIdentifier)
-        return applicationSupportDirectoryURL
-            .appendingPathComponent(namespace, isDirectory: true)
-            .appendingPathComponent("vscode-serve-web", isDirectory: true)
-    }
-
-    private static func resolvePort(
-        environment: [String: String],
-        defaults: UserDefaults,
-        bundleIdentifier: String?
-    ) -> (port: Int, allowsEphemeralFallback: Bool) {
-        if let rawPort = environment[portEnvironmentKey],
-           let port = Int(rawPort),
-           isValidPort(port) {
-            return (port, false)
-        }
-
-        let storedPort = defaults.integer(forKey: portDefaultsKey)
-        if isValidPort(storedPort) {
-            return (storedPort, true)
-        }
-
-        let port = defaultStablePort(bundleIdentifier)
-        defaults.set(port, forKey: portDefaultsKey)
-        return (port, true)
-    }
-
-    private static func defaultStablePort(_ bundleIdentifier: String?) -> Int {
-        let identifier = bundleIdentifier?.isEmpty == false ? bundleIdentifier! : "cmux"
-        let dynamicPortStart = 49152
-        let dynamicPortCount = 16384
-        let offset = identifier.unicodeScalars.reduce(0) { partial, scalar in
-            (partial &* 31 &+ Int(scalar.value)) % dynamicPortCount
-        }
-        return dynamicPortStart + offset
-    }
-
-    private static func isValidPort(_ port: Int) -> Bool {
-        port > 0 && port <= 65535
-    }
-
-    private static func ensureConnectionTokenFile(
-        in serverDataDirectoryURL: URL,
-        fileManager: FileManager
-    ) -> URL? {
-        let tokenFileURL = serverDataDirectoryURL.appendingPathComponent(
-            "connection-token",
-            isDirectory: false
-        )
-        if fileManager.fileExists(atPath: tokenFileURL.path) {
-            if isUsableConnectionTokenFile(tokenFileURL, fileManager: fileManager) {
-                return tokenFileURL
-            }
-            try? fileManager.removeItem(at: tokenFileURL)
-        }
-
-        let token = randomConnectionToken()
-        guard let tokenData = token.data(using: .utf8) else { return nil }
-
-        let fileDescriptor = open(tokenFileURL.path, O_WRONLY | O_CREAT | O_EXCL, S_IRUSR | S_IWUSR)
-        if fileDescriptor < 0 {
-            return fileManager.fileExists(atPath: tokenFileURL.path)
-                && isUsableConnectionTokenFile(tokenFileURL, fileManager: fileManager)
-                ? tokenFileURL
-                : nil
-        }
-        defer { _ = close(fileDescriptor) }
-
-        let wroteAllBytes = tokenData.withUnsafeBytes { rawBuffer in
-            guard let baseAddress = rawBuffer.baseAddress else { return false }
-            return write(fileDescriptor, baseAddress, rawBuffer.count) == rawBuffer.count
-        }
-        guard wroteAllBytes else {
-            try? fileManager.removeItem(at: tokenFileURL)
-            return nil
-        }
-
-        return tokenFileURL
-    }
-
-    private static func randomConnectionToken() -> String {
-        UUID().uuidString.replacingOccurrences(of: "-", with: "")
-    }
-
-    private static func isUsableConnectionTokenFile(
-        _ tokenFileURL: URL,
-        fileManager: FileManager
-    ) -> Bool {
-        guard let attributes = try? fileManager.attributesOfItem(atPath: tokenFileURL.path),
-              let fileSize = attributes[.size] as? NSNumber,
-              fileSize.intValue == 32,
-              let permissions = attributes[.posixPermissions] as? NSNumber,
-              permissions.intValue & 0o777 == 0o600,
-              let data = try? Data(contentsOf: tokenFileURL),
-              let token = String(data: data, encoding: .utf8) else {
-            return false
-        }
-        return token.range(of: #"^[0-9A-Fa-f]{32}$"#, options: .regularExpression) != nil
-    }
-
-    private static func expandedHomePath(_ path: String) -> String {
-        guard path == "~" || path.hasPrefix("~/") else { return path }
-        let homePath = FileManager.default.homeDirectoryForCurrentUser.path
-        if path == "~" {
-            return homePath
-        }
-        return homePath + String(path.dropFirst())
-    }
-
-    private static func normalizedNamespace(_ bundleIdentifier: String?) -> String {
-        guard let bundleIdentifier,
-              !bundleIdentifier.isEmpty else {
-            return "cmux"
-        }
-        return bundleIdentifier
-            .replacingOccurrences(of: "[^A-Za-z0-9.-]", with: "-", options: .regularExpression)
-    }
-}
-
-enum VSCodeCLILaunchConfigurationBuilder {
-    private struct VSCodeProductMetadata: Decodable {
-        let dataFolderName: String?
-    }
-
-    static func launchConfiguration(
-        vscodeApplicationURL: URL,
-        homeDirectoryURL: URL = FileManager.default.homeDirectoryForCurrentUser,
-        baseEnvironment: [String: String] = ProcessInfo.processInfo.environment,
-        isExecutableAtPath: (String) -> Bool = { FileManager.default.isExecutableFile(atPath: $0) },
-        dataAtURL: (URL) -> Data? = { try? Data(contentsOf: $0) },
-        contentsOfDirectoryAtURL: (URL) -> [URL] = { url in
-            (try? FileManager.default.contentsOfDirectory(
-                at: url,
-                includingPropertiesForKeys: [.contentModificationDateKey],
-                options: [.skipsHiddenFiles]
-            )) ?? []
-        },
-        contentModificationDateAtURL: (URL) -> Date? = { url in
-            (try? url.resourceValues(forKeys: [.contentModificationDateKey]))?.contentModificationDate
-        }
-    ) -> VSCodeCLILaunchConfiguration? {
-        let contentsURL = vscodeApplicationURL.appendingPathComponent("Contents", isDirectory: true)
-        let environment = nodeSafeEnvironment(from: baseEnvironment)
-        let codeTunnelURL = contentsURL.appendingPathComponent("Resources/app/bin/code-tunnel", isDirectory: false)
-
-        if isExecutableAtPath(codeTunnelURL.path) {
-            var codeTunnelEnvironment = environment
-            codeTunnelEnvironment["ELECTRON_RUN_AS_NODE"] = "1"
-            return VSCodeCLILaunchConfiguration(
-                executableURL: codeTunnelURL,
-                argumentsPrefix: ["serve-web"],
-                environment: codeTunnelEnvironment,
-                usesCodeTunnelWrapper: true,
-                supportsUserDataDirectoryArgument: false
-            )
-        }
-
-        if let codeServerURL = preferredCachedCodeServerURL(
-            contentsURL: contentsURL,
-            homeDirectoryURL: homeDirectoryURL,
-            isExecutableAtPath: isExecutableAtPath,
-            dataAtURL: dataAtURL,
-            contentsOfDirectoryAtURL: contentsOfDirectoryAtURL,
-            contentModificationDateAtURL: contentModificationDateAtURL
-        ) {
-            var codeServerEnvironment = environment
-            codeServerEnvironment.removeValue(forKey: "ELECTRON_RUN_AS_NODE")
-            return VSCodeCLILaunchConfiguration(
-                executableURL: codeServerURL,
-                argumentsPrefix: [],
-                environment: codeServerEnvironment,
-                usesCodeTunnelWrapper: false,
-                supportsUserDataDirectoryArgument: true
-            )
-        }
-
-        return nil
-    }
-
-    private static func nodeSafeEnvironment(from baseEnvironment: [String: String]) -> [String: String] {
-        var environment = baseEnvironment
-        environment.removeValue(forKey: "VSCODE_NODE_OPTIONS")
-        environment.removeValue(forKey: "VSCODE_NODE_REPL_EXTERNAL_MODULE")
-        if let nodeOptions = environment["NODE_OPTIONS"] {
-            environment["VSCODE_NODE_OPTIONS"] = nodeOptions
-        }
-        if let nodeReplExternalModule = environment["NODE_REPL_EXTERNAL_MODULE"] {
-            environment["VSCODE_NODE_REPL_EXTERNAL_MODULE"] = nodeReplExternalModule
-        }
-        environment.removeValue(forKey: "NODE_OPTIONS")
-        environment.removeValue(forKey: "NODE_REPL_EXTERNAL_MODULE")
-        return environment
-    }
-
-    private static func preferredCachedCodeServerURL(
-        contentsURL: URL,
-        homeDirectoryURL: URL,
-        isExecutableAtPath: (String) -> Bool,
-        dataAtURL: (URL) -> Data?,
-        contentsOfDirectoryAtURL: (URL) -> [URL],
-        contentModificationDateAtURL: (URL) -> Date?
-    ) -> URL? {
-        let dataFolderName = vscodeDataFolderName(
-            contentsURL: contentsURL,
-            dataAtURL: dataAtURL
-        )
-        let serveWebCacheURL = homeDirectoryURL
-            .appendingPathComponent(dataFolderName, isDirectory: true)
-            .appendingPathComponent("cli/serve-web", isDirectory: true)
-
-        if let orderedCacheIDs = serveWebLRUCacheIDs(
-            serveWebCacheURL: serveWebCacheURL,
-            dataAtURL: dataAtURL
-        ) {
-            for cacheID in orderedCacheIDs {
-                let codeServerURL = serveWebCacheURL
-                    .appendingPathComponent(cacheID, isDirectory: true)
-                    .appendingPathComponent("bin/code-server", isDirectory: false)
-                if isExecutableAtPath(codeServerURL.path) {
-                    return codeServerURL
-                }
-            }
-        }
-
-        let candidates = contentsOfDirectoryAtURL(serveWebCacheURL)
-            .map {
-                $0.appendingPathComponent("bin/code-server", isDirectory: false)
-            }
-            .filter {
-                isExecutableAtPath($0.path)
-            }
-            .sorted { lhs, rhs in
-                let lhsDate = contentModificationDateAtURL(lhs) ?? .distantPast
-                let rhsDate = contentModificationDateAtURL(rhs) ?? .distantPast
-                if lhsDate != rhsDate {
-                    return lhsDate > rhsDate
-                }
-                return lhs.path > rhs.path
-            }
-
-        return candidates.first
-    }
-
-    private static func vscodeDataFolderName(
-        contentsURL: URL,
-        dataAtURL: (URL) -> Data?
-    ) -> String {
-        let productURL = contentsURL.appendingPathComponent("Resources/app/product.json", isDirectory: false)
-        guard let data = dataAtURL(productURL),
-              let product = try? JSONDecoder().decode(VSCodeProductMetadata.self, from: data),
-              let dataFolderName = product.dataFolderName,
-              isSafePathComponent(dataFolderName) else {
-            return ".vscode"
-        }
-        return dataFolderName
-    }
-
-    private static func serveWebLRUCacheIDs(
-        serveWebCacheURL: URL,
-        dataAtURL: (URL) -> Data?
-    ) -> [String]? {
-        let lruURL = serveWebCacheURL.appendingPathComponent("lru.json", isDirectory: false)
-        guard let data = dataAtURL(lruURL),
-              let cacheIDs = try? JSONDecoder().decode([String].self, from: data) else {
-            return nil
-        }
-        return cacheIDs.filter(isSafePathComponent)
-    }
-
-    private static func isSafePathComponent(_ component: String) -> Bool {
-        guard !component.isEmpty, component != ".", component != ".." else { return false }
-        return component.rangeOfCharacter(from: CharacterSet(charactersIn: "/\\")) == nil
-    }
-}
-
-final class VSCodeServeWebController {
-    static let shared = VSCodeServeWebController()
-    private static let serveWebStartupTimeoutSeconds: TimeInterval = 60
-
-    private let queue = DispatchQueue(label: "cmux.vscode.serveWeb")
-    private let launchQueue = DispatchQueue(label: "cmux.vscode.serveWeb.launch")
-    private let launchProcessOverride: ((URL, UInt64) -> (process: Process, url: URL)?)?
-    private var serveWebProcess: Process?
-    private var launchingProcess: Process?
-    private var serveWebURL: URL?
-    private var pendingCompletions: [(generation: UInt64, completion: (URL?) -> Void)] = []
-    private var isLaunching = false
-    private var activeLaunchGeneration: UInt64?
-    private var lifecycleGeneration: UInt64 = 0
-
-    init(launchProcessOverride: ((URL, UInt64) -> (process: Process, url: URL)?)? = nil) {
-        self.launchProcessOverride = launchProcessOverride
-    }
-
-    func ensureServeWebURL(vscodeApplicationURL: URL, completion: @escaping (URL?) -> Void) {
-        queue.async {
-            if let process = self.serveWebProcess,
-               process.isRunning,
-               let url = self.serveWebURL {
-                DispatchQueue.main.async {
-                    completion(url)
-                }
-                return
-            }
-
-            let completionGeneration = self.lifecycleGeneration
-            self.pendingCompletions.append((generation: completionGeneration, completion: completion))
-            guard !self.isLaunching else { return }
-
-            self.isLaunching = true
-            let launchGeneration = completionGeneration
-            self.activeLaunchGeneration = launchGeneration
-
-            self.launchQueue.async {
-                let shouldLaunch = self.queue.sync {
-                    self.lifecycleGeneration == launchGeneration
-                }
-                guard shouldLaunch else {
-                    self.queue.async {
-                        guard self.activeLaunchGeneration == launchGeneration else { return }
-                        self.isLaunching = false
-                        self.activeLaunchGeneration = nil
-                    }
-                    return
-                }
-                let launchResult = self.launchServeWebProcess(
-                    vscodeApplicationURL: vscodeApplicationURL,
-                    expectedGeneration: launchGeneration
-                )
-                self.queue.async {
-                    guard self.activeLaunchGeneration == launchGeneration else {
-                        if let process = launchResult?.process, process.isRunning {
-                            process.terminate()
-                        }
-                        return
-                    }
-                    self.isLaunching = false
-                    self.activeLaunchGeneration = nil
-
-                    guard self.lifecycleGeneration == launchGeneration else {
-                        if let launchedProcess = launchResult?.process,
-                           self.launchingProcess === launchedProcess {
-                            self.launchingProcess = nil
-                        }
-                        if let process = launchResult?.process, process.isRunning {
-                            process.terminate()
-                        }
-                        return
-                    }
-
-                    if let launchResult {
-                        self.launchingProcess = nil
-                        self.serveWebProcess = launchResult.process
-                        self.serveWebURL = launchResult.url
-                    } else {
-                        self.launchingProcess = nil
-                        self.serveWebProcess = nil
-                        self.serveWebURL = nil
-                    }
-
-                    var completions: [(URL?) -> Void] = []
-                    var remaining: [(generation: UInt64, completion: (URL?) -> Void)] = []
-                    for pending in self.pendingCompletions {
-                        if pending.generation == launchGeneration {
-                            completions.append(pending.completion)
-                        } else {
-                            remaining.append(pending)
-                        }
-                    }
-                    self.pendingCompletions = remaining
-                    let resolvedURL = self.serveWebURL
-                    DispatchQueue.main.async {
-                        completions.forEach { $0(resolvedURL) }
-                    }
-                }
-            }
-        }
-    }
-
-    func stop() {
-        let (processes, completions): ([Process], [(URL?) -> Void]) = queue.sync {
-            self.lifecycleGeneration &+= 1
-            self.isLaunching = false
-            self.activeLaunchGeneration = nil
-            var processes: [Process] = []
-            if let process = self.serveWebProcess {
-                processes.append(process)
-            }
-            if let process = self.launchingProcess,
-               !processes.contains(where: { $0 === process }) {
-                processes.append(process)
-            }
-            self.serveWebProcess = nil
-            self.launchingProcess = nil
-            self.serveWebURL = nil
-            let completions = self.pendingCompletions.map(\.completion)
-            self.pendingCompletions.removeAll()
-            return (processes, completions)
-        }
-
-        for process in processes where process.isRunning {
-            process.terminate()
-        }
-
-        if !completions.isEmpty {
-            DispatchQueue.main.async {
-                completions.forEach { $0(nil) }
-            }
-        }
-    }
-
-    func restart(vscodeApplicationURL: URL, completion: @escaping (URL?) -> Void) {
-        stop()
-        ensureServeWebURL(vscodeApplicationURL: vscodeApplicationURL, completion: completion)
-    }
-
-    func isServeWebURL(_ candidateURL: URL?) -> Bool {
-        guard let candidateURL else { return false }
-        let serveWebURL = queue.sync {
-            self.serveWebURL
-        }
-        return Self.urlsShareLoopbackOrigin(candidateURL, serveWebURL)
-    }
-
-    private func launchServeWebProcess(
-        vscodeApplicationURL: URL,
-        expectedGeneration: UInt64
-    ) -> (process: Process, url: URL)? {
-        if let launchProcessOverride {
-            return launchProcessOverride(vscodeApplicationURL, expectedGeneration)
-        }
-
-        guard let launchConfiguration = VSCodeCLILaunchConfigurationBuilder.launchConfiguration(
-            vscodeApplicationURL: vscodeApplicationURL
-        ) else { return nil }
-
-        // Use the child-process environment so CMUX_* overrides that survive
-        // nodeSafeEnvironment are also honored by serve-web option resolution.
-        guard let launchOptions = VSCodeServeWebLaunchOptions.resolve(
-            environment: launchConfiguration.environment
-        ) else { return nil }
-
-        let launchOptionCandidates = [launchOptions] + [launchOptions.ephemeralPortFallback()].compactMap { $0 }
-        for launchOptions in launchOptionCandidates {
-            if let launchResult = launchServeWebProcess(
-                launchConfiguration: launchConfiguration,
-                launchOptions: launchOptions,
-                expectedGeneration: expectedGeneration
-            ) {
-                return launchResult
-            }
-        }
-        return nil
-    }
-
-    private func launchServeWebProcess(
-        launchConfiguration: VSCodeCLILaunchConfiguration,
-        launchOptions: VSCodeServeWebLaunchOptions,
-        expectedGeneration: UInt64
-    ) -> (process: Process, url: URL)? {
-        let process = Process()
-        process.executableURL = launchConfiguration.executableURL
-        process.arguments = launchConfiguration.processArguments(for: launchOptions)
-        process.environment = launchConfiguration.processEnvironment(for: launchOptions)
-
-        let stdoutPipe = Pipe()
-        let stderrPipe = Pipe()
-        process.standardOutput = stdoutPipe
-        process.standardError = stderrPipe
-
-        let collector = ServeWebOutputCollector()
-        let outputReader: (FileHandle) -> Void = { fileHandle in
-            switch fileHandle.readAvailableDataOrEndOfFile() {
-            case .data(let data):
-                collector.append(data)
-            case .wouldBlock:
-                return
-            case .endOfFile:
-                fileHandle.readabilityHandler = nil
-            }
-        }
-        stdoutPipe.fileHandleForReading.readabilityHandler = outputReader
-        stderrPipe.fileHandleForReading.readabilityHandler = outputReader
-
-        process.terminationHandler = { [weak self] terminatedProcess in
-            stdoutPipe.fileHandleForReading.readabilityHandler = nil
-            stderrPipe.fileHandleForReading.readabilityHandler = nil
-            Self.drainAvailableOutput(from: stdoutPipe.fileHandleForReading, collector: collector)
-            Self.drainAvailableOutput(from: stderrPipe.fileHandleForReading, collector: collector)
-            collector.markProcessExited()
-            self?.queue.async {
-                guard let self else { return }
-                if self.launchingProcess === terminatedProcess {
-                    self.launchingProcess = nil
-                }
-                if self.serveWebProcess === terminatedProcess {
-                    self.serveWebProcess = nil
-                    self.serveWebURL = nil
-                }
-            }
-        }
-
-        let didStart: Bool = queue.sync {
-            guard self.lifecycleGeneration == expectedGeneration,
-                  self.activeLaunchGeneration == expectedGeneration else {
-                return false
-            }
-            self.launchingProcess = process
-            do {
-                try process.run()
-                return true
-            } catch {
-                if self.launchingProcess === process {
-                    self.launchingProcess = nil
-                }
-                return false
-            }
-        }
-        guard didStart else {
-            stdoutPipe.fileHandleForReading.readabilityHandler = nil
-            stderrPipe.fileHandleForReading.readabilityHandler = nil
-            return nil
-        }
-
-        guard collector.waitForURL(timeoutSeconds: Self.serveWebStartupTimeoutSeconds),
-              let serveWebURL = collector.webUIURL else {
-            stdoutPipe.fileHandleForReading.readabilityHandler = nil
-            stderrPipe.fileHandleForReading.readabilityHandler = nil
-            if process.isRunning {
-                process.terminate()
-            } else {
-                queue.sync {
-                    if self.launchingProcess === process {
-                        self.launchingProcess = nil
-                    }
-                    if self.serveWebProcess === process {
-                        self.serveWebProcess = nil
-                        self.serveWebURL = nil
-                    }
-                }
-            }
-            return nil
-        }
-
-        return (process, serveWebURL)
-    }
-
-    private static func drainAvailableOutput(from fileHandle: FileHandle, collector: ServeWebOutputCollector) {
-        while true {
-            switch fileHandle.readAvailableDataOrEndOfFile() {
-            case .data(let data):
-                collector.append(data)
-            case .wouldBlock, .endOfFile:
-                return
-            }
-        }
-    }
-
-    private static func urlsShareLoopbackOrigin(_ lhs: URL, _ rhs: URL?) -> Bool {
-        guard let rhs else { return false }
-        guard lhs.scheme?.lowercased() == "http",
-              rhs.scheme?.lowercased() == "http" else {
-            return false
-        }
-        guard lhs.port == rhs.port, lhs.port != nil else { return false }
-        guard let lhsHost = BrowserInsecureHTTPSettings.normalizeHost(lhs.host ?? ""),
-              let rhsHost = BrowserInsecureHTTPSettings.normalizeHost(rhs.host ?? "") else {
-            return false
-        }
-        return RemoteLoopbackProxyAlias.isLoopbackHost(lhsHost)
-            && RemoteLoopbackProxyAlias.isLoopbackHost(rhsHost)
-    }
-}
-
-final class ServeWebOutputCollector {
-    private let lock = NSLock()
-    private let semaphore = DispatchSemaphore(value: 0)
-    private var outputBuffer = ""
-    private var resolvedURL: URL?
-    private var didSignal = false
-
-    var webUIURL: URL? {
-        lock.lock()
-        defer { lock.unlock() }
-        return resolvedURL
-    }
-
-    func append(_ data: Data) {
-        guard let text = String(data: data, encoding: .utf8), !text.isEmpty else { return }
-        lock.lock()
-        defer { lock.unlock() }
-        guard resolvedURL == nil else { return }
-        outputBuffer.append(text)
-        while let newlineIndex = outputBuffer.firstIndex(where: \.isNewline) {
-            let line = String(outputBuffer[..<newlineIndex])
-            outputBuffer.removeSubrange(...newlineIndex)
-            guard let parsedURL = VSCodeServeWebURLBuilder.extractWebUIURL(from: line) else {
-                continue
-            }
-            resolvedURL = parsedURL
-            outputBuffer.removeAll(keepingCapacity: false)
-            if !didSignal {
-                didSignal = true
-                semaphore.signal()
-            }
-            return
-        }
-    }
-
-    func markProcessExited() {
-        lock.lock()
-        defer { lock.unlock() }
-        if resolvedURL == nil, !outputBuffer.isEmpty,
-           let parsedURL = VSCodeServeWebURLBuilder.extractWebUIURL(from: outputBuffer) {
-            resolvedURL = parsedURL
-            outputBuffer.removeAll(keepingCapacity: false)
-        }
-        guard !didSignal else { return }
-        didSignal = true
-        semaphore.signal()
-    }
-
-    func waitForURL(timeoutSeconds: TimeInterval) -> Bool {
-        if webUIURL != nil { return true }
-        _ = semaphore.wait(timeout: .now() + timeoutSeconds)
-        return webUIURL != nil
     }
 }
