+++ Sources/App/VSCodeCLILaunchConfigurationBuilder.swift
@@ -1,0 +1,157 @@
+import Foundation
+
+struct VSCodeCLILaunchConfigurationBuilder {
+    private struct VSCodeProductMetadata: Decodable {
+        let dataFolderName: String?
+    }
+
+    func launchConfiguration(
+        vscodeApplicationURL: URL,
+        homeDirectoryURL: URL = FileManager.default.homeDirectoryForCurrentUser,
+        baseEnvironment: [String: String] = ProcessInfo.processInfo.environment,
+        isExecutableAtPath: (String) -> Bool = { FileManager.default.isExecutableFile(atPath: $0) },
+        dataAtURL: (URL) -> Data? = { try? Data(contentsOf: $0) },
+        contentsOfDirectoryAtURL: (URL) -> [URL] = { url in
+            (try? FileManager.default.contentsOfDirectory(
+                at: url,
+                includingPropertiesForKeys: [.contentModificationDateKey],
+                options: [.skipsHiddenFiles]
+            )) ?? []
+        },
+        contentModificationDateAtURL: (URL) -> Date? = { url in
+            (try? url.resourceValues(forKeys: [.contentModificationDateKey]))?.contentModificationDate
+        }
+    ) -> VSCodeCLILaunchConfiguration? {
+        let contentsURL = vscodeApplicationURL.appendingPathComponent("Contents", isDirectory: true)
+        let environment = nodeSafeEnvironment(from: baseEnvironment)
+        let codeTunnelURL = contentsURL.appendingPathComponent("Resources/app/bin/code-tunnel", isDirectory: false)
+
+        if isExecutableAtPath(codeTunnelURL.path) {
+            var codeTunnelEnvironment = environment
+            codeTunnelEnvironment["ELECTRON_RUN_AS_NODE"] = "1"
+            return VSCodeCLILaunchConfiguration(
+                executableURL: codeTunnelURL,
+                argumentsPrefix: ["serve-web"],
+                environment: codeTunnelEnvironment,
+                usesCodeTunnelWrapper: true,
+                supportsUserDataDirectoryArgument: false
+            )
+        }
+
+        if let codeServerURL = preferredCachedCodeServerURL(
+            contentsURL: contentsURL,
+            homeDirectoryURL: homeDirectoryURL,
+            isExecutableAtPath: isExecutableAtPath,
+            dataAtURL: dataAtURL,
+            contentsOfDirectoryAtURL: contentsOfDirectoryAtURL,
+            contentModificationDateAtURL: contentModificationDateAtURL
+        ) {
+            var codeServerEnvironment = environment
+            codeServerEnvironment.removeValue(forKey: "ELECTRON_RUN_AS_NODE")
+            return VSCodeCLILaunchConfiguration(
+                executableURL: codeServerURL,
+                argumentsPrefix: [],
+                environment: codeServerEnvironment,
+                usesCodeTunnelWrapper: false,
+                supportsUserDataDirectoryArgument: true
+            )
+        }
+
+        return nil
+    }
+
+    private func nodeSafeEnvironment(from baseEnvironment: [String: String]) -> [String: String] {
+        var environment = baseEnvironment
+        environment.removeValue(forKey: "VSCODE_NODE_OPTIONS")
+        environment.removeValue(forKey: "VSCODE_NODE_REPL_EXTERNAL_MODULE")
+        if let nodeOptions = environment["NODE_OPTIONS"] {
+            environment["VSCODE_NODE_OPTIONS"] = nodeOptions
+        }
+        if let nodeReplExternalModule = environment["NODE_REPL_EXTERNAL_MODULE"] {
+            environment["VSCODE_NODE_REPL_EXTERNAL_MODULE"] = nodeReplExternalModule
+        }
+        environment.removeValue(forKey: "NODE_OPTIONS")
+        environment.removeValue(forKey: "NODE_REPL_EXTERNAL_MODULE")
+        return environment
+    }
+
+    private func preferredCachedCodeServerURL(
+        contentsURL: URL,
+        homeDirectoryURL: URL,
+        isExecutableAtPath: (String) -> Bool,
+        dataAtURL: (URL) -> Data?,
+        contentsOfDirectoryAtURL: (URL) -> [URL],
+        contentModificationDateAtURL: (URL) -> Date?
+    ) -> URL? {
+        let dataFolderName = vscodeDataFolderName(
+            contentsURL: contentsURL,
+            dataAtURL: dataAtURL
+        )
+        let serveWebCacheURL = homeDirectoryURL
+            .appendingPathComponent(dataFolderName, isDirectory: true)
+            .appendingPathComponent("cli/serve-web", isDirectory: true)
+
+        if let orderedCacheIDs = serveWebLRUCacheIDs(
+            serveWebCacheURL: serveWebCacheURL,
+            dataAtURL: dataAtURL
+        ) {
+            for cacheID in orderedCacheIDs {
+                let codeServerURL = serveWebCacheURL
+                    .appendingPathComponent(cacheID, isDirectory: true)
+                    .appendingPathComponent("bin/code-server", isDirectory: false)
+                if isExecutableAtPath(codeServerURL.path) {
+                    return codeServerURL
+                }
+            }
+        }
+
+        let candidates = contentsOfDirectoryAtURL(serveWebCacheURL)
+            .map {
+                $0.appendingPathComponent("bin/code-server", isDirectory: false)
+            }
+            .filter {
+                isExecutableAtPath($0.path)
+            }
+            .sorted { lhs, rhs in
+                let lhsDate = contentModificationDateAtURL(lhs) ?? .distantPast
+                let rhsDate = contentModificationDateAtURL(rhs) ?? .distantPast
+                if lhsDate != rhsDate {
+                    return lhsDate > rhsDate
+                }
+                return lhs.path > rhs.path
+            }
+
+        return candidates.first
+    }
+
+    private func vscodeDataFolderName(
+        contentsURL: URL,
+        dataAtURL: (URL) -> Data?
+    ) -> String {
+        let productURL = contentsURL.appendingPathComponent("Resources/app/product.json", isDirectory: false)
+        guard let data = dataAtURL(productURL),
+              let product = try? JSONDecoder().decode(VSCodeProductMetadata.self, from: data),
+              let dataFolderName = product.dataFolderName,
+              isSafePathComponent(dataFolderName) else {
+            return ".vscode"
+        }
+        return dataFolderName
+    }
+
+    private func serveWebLRUCacheIDs(
+        serveWebCacheURL: URL,
+        dataAtURL: (URL) -> Data?
+    ) -> [String]? {
+        let lruURL = serveWebCacheURL.appendingPathComponent("lru.json", isDirectory: false)
+        guard let data = dataAtURL(lruURL),
+              let cacheIDs = try? JSONDecoder().decode([String].self, from: data) else {
+            return nil
+        }
+        return cacheIDs.filter(isSafePathComponent)
+    }
+
+    private func isSafePathComponent(_ component: String) -> Bool {
+        guard !component.isEmpty, component != ".", component != ".." else { return false }
+        return component.rangeOfCharacter(from: CharacterSet(charactersIn: "/\\")) == nil
+    }
+}
